Remove Max layout, increase window borders in all tests.
The Max layout is equivalent to a Stack with a 0-pixel border and a single
column. None the less, it is worth while having a Max implementation (as an
example of a simple layout), but it should be re-implemented so that it tracks
window order itself, rather than using the inherent ordering of the group to
which it is attached - expecially since this ordering is going away in the next
commit.

The window border increase in the test suite looks ugly but makes all window
boundaries clear.