Increased error tolerance when dealing with nonexistent keysyms.
The keysym_to_keycode method will raise a KeyError when handling an
out-of-range keysym. In my setup, this happens when I try to establish
a remote session from OS X.
I'm not sure about the why, but I think it's a good idea to make qtile
more tolerant to that kind of errors.

I added a bounds-check that returns zero when it encounters a
nonexistent keysym.