Fixed a bug that misinterpreted the display name in remote sessions.
The display number must always be two-part for the IPC class to work
correctly (:0.0 instead of :0). The check in manager.py appends
a ".0" if it can't find a dot in the display name.
However, there might be a dot in the display name that is not part
of the display number, e.g. during remote sessions where the hostname
contains a dot (e.g. displayName == "host.name.net:0").

I extended the check for the correct format of displayName to strip
the host part first.