Give widgets the ability to receive keyboard input.
This is needed for the minibuffer widget. Any widget can now gain
keyboard focus by calling a method of its bar object. It will retain
focus until it calls the corresponding ungrab method (This could be a
problem if a widget forgets to do that)

I had to modify the eventMask of all internal windows to receive
KeyPress events. I don't know if this has performance implications
or other side effects.