A widget to receive keyboard input from the user.
The minibuffer widget can be used for any kind of user input.
First, it must be configured as a widget in the config file. Then,
we can call its startInput method, which will cause the widget to
display a prompt, acquire keyboard focus and take one line of user
input until either Return or Escape is pressed.
After that, a callback method is called with the input string as
argument.

Due to the lack of a LookupString function in XCB, the widget can
currently only handle ASCII characters, as it simply takes any
keysym < 127 as an ASCII code.