fix window mapping on restart bug
When X-Server restarts, mapped windows shpuld be re-mapped, which
they weren't. The scan function (in manager.py) that should of
been doing this, was using a window attribute to determine which
windows should be remapped. But it seems thatthis attribute would
be set to a value indicating 'non-visible' for all but the window
in focus at the time of the server restart; windows in other
groups, that is. Instead, it now uses the WM hint, 'initial_state'
to determine which windows should appear. Please see the
comments in 'scan' as to if any imporovements are needed; I think
it might not be necessary to check map_state anymore. Also the
window in focus before and after the restart might change, maybe
map_state can be useful afterall? Finally, feel free to complain
to me if you think this message is too verbose :>