Replaced multichar with struct.pack in xcbq. This fixes a bug with opacity.
utils.multichar() is like struct.pack(), but always big-endian. On systems
where this isn't the native endianness, long int window properties such as
opacity were set with wrong values.

With this change, utils.multichar() is no longer used. There is also
ipc.multichar() which is the same, but it's used with width=3, and can't
be replaced with struct. I suggest changing the HDRLEN variable to 4 and
using struct.pack("L", ...), but I'll leave this to a core developer.