Refine explicit layout setting on the Group object.
- Remove functionality to set a layout using a Layout object. At the moment,
all groups have the same set of layouts - this breaks that expectation, by
letting the user add a new layout to a group, not shared with all other groups.
I think that letting each group have a different set of layouts is probably a
desirable feature, but I'd like to think more about how to do this in a
user-friendly way.

- Add a "setlayout" command, using the layout setter.

- Add unit tests.