More graph work.
- Add a linefill graph type.

- Make the bounding rectangle a hard rectangle, rather than a rounded one.
Using a rounded rect, the graph "bleeds out" on the rounded corners, making it
look un-neat. If someone wanted to add the rounded rect back in, I wouldn't
object, but we'd have to do the work to take the corner rounding radius into
account, and fit the graph completely "within" the rectangle.