Add all managed windows to the save-set
Windows in the save-set are not only reparented to root, but also
mapped, which is exactly what we need when switching to a different
window manager. So yes, this fixes the bug with missing windows when
replacing qtile with awesome (and other window managers probably)

Thanks to Uli Schlachter who helped me on irc to get this working.