fixed qt aplications never come back from close to tray
For compatibility with obsolete clients, window managers should trigger
the transition to the Withdrawn state on the real UnmapNotify rather
than waiting for the synthetic one. They should also trigger the
transition if they receive a synthetic UnmapNotify on a window for which
they have not yet received a real UnmapNotify.
http://tronche.com/gui/x/icccm/sec-4.html#s-4.1.4