Refactor config file reading a little more
This commit unbreaks the unit tests (there was a ciruclar import problem with
some of them, so it just breaks the cycle). In additon, it eliminates
confreader.Config entirely, using resources/default_config.py as the default
config if the user specified no config file (or we couldn't find the one they
did specify).

To add new config options:
  * add the name of your option to libqtile/confreader.py
  * add a reasonable default in libqtile/resources/default_config.py