Adds reconnect option to Mpd-widget
If reconnect is True the widget will keep trying to reconnect to the mpd
daemon. This is useful if mpd isn't started when logging into X, but
will be started manually by the user later.

If reconnect is False the default behavior will be used (the widget
will try to connect one time and not try anymore.).