Completed threading of the GoogleCalendar widget
isolated run(FLOW, storage) in a separate thread, so that it
doesn't block qtile during the authentication process. This
required removing one line from the oauth2client.tools module,
specifically storage.put(), as that line was not thread-safe.
The revised module is included in the qtile/libqtile directory,
and is called custom_tools.py. As a result of this modification,
we have to remember to storage.put() our credentials ourselves
in the main qtile thread.