Update pacman.py
Checkupdates is a non-sudo way of checking for updates as using pacman -Qu only checks for out-of-date packages with your local package list. Pacman -Su updates the package lists but requires sudo access to do so. 

You could have a cron job that updates the package lists and then pacman -Qu would suffice.