Preventing NotificationService from registering if unused
Until now, once libqtile.widget was imported, a new NotificationService was
created and registered as a DBus Notification daemon, and thus making
impossible for any other notification daemon to register.

This patch prevents NotificationManager.service (a NotificationService
instance) to be initialized until further call of register() method, which
basically avoid registering as a notification daemon when the configuration
does not uses any notification widget.

Closes #335