Set initial widget sizes on bar _configure()
With the Bar queue fix in cde3ed395, it is possible for a widget to ask to
draw() itself before the bar has drawn the first time (e.g. if someone does a
direct add to the event loop queue or sets a short timeout [graph widget]). I
consider the offset/width to be part of a widget's configuration, and thus they
should really be set in _configure() anyway, so we do that here, avoiding the
potential race.

Closes #347