Pseudoscreens with same x and y share Screen
Pseudoscreens with the same x and y coordinate will not be assigned
their different screens. Rather, a single screen will be created with
the largest width and height of the pseudoscreens sharing the same
origin. This resolves issues when cloning displays, as both Screens
would occupy the same location, and maneuvering groups in this
configuration was tricky.

Resolves #378