Revert transparency stuff in max
This reverts commit 874fb678e71e64112d68ac8ae0ab944c11817c37.
This reverts commit 1b46a8b8f70a46f3a0bcc1cb364933db8de42159.

Closes #419
Closes #421
Closes #423
Reopens #415

The problem with this approach is that it abuses X's "above" semantics, causing
lots of odd issues with floating windows (viz. the issues above). The real
solution here is to implement proper Z axis knowledge within qtile and then set
it accordingly. Based on some discussions in IRC, we decided it was best to
revert this and re-open the feature request and do a better implementation in
the future.