Don't crash on widget errors
We could return "" here, but it seems more useful to let users know that a
particular widget is having problems. We can change this to "" if it becomes
more of an issue. Whatever we do, we can't return None, because we immediately
try to parse None for pango markup and fail.