fix a hang on qtile shutdown when using gobject
Since qtile started a thread for running the gobject event loop and blocks in
that thread until it sees an event, shutting down qtile hangs until it sees
another event (e.g. a dbus event). Here, we just fake an event so that the call
to iteration() returns and that thread sees that the asyncio event loop is
dead, so it exits.