Fix having long unicode characters in properties
The real problem here is that for long unicode characters in a string,
the number of bytes needed is more than the number of characters in the
string.  By encoding unicode before taking the length, we mitigate this
problem.

When we have things that are unicode (PY3 string or PY2 unicode), we
really _should_ encode them properly before sending them over the wire.
xcffib does have code that could deal with strings if we were able to
properly get the length, but we should play nice and send it bytes.