work around AttributeError from asyncio
After a bunch of research and staring at the code for a while, I can't see how
this would happen except in older versions of asyncio. However, I can't find
much discussion of this particular race anywhere on the web (indeed, it
happens only very rarely on travis even for us, and I've never seen it in the
wild).

https://github.com/aaugustin/websockets/issues/23 is the only other discussion
of this issue that I can find, and they've solved it in a similar way.