allow pypy build to fail
also re-add pypy3 build, and allow it to fail too. It seems that these two are
fickle, but the rest of the build seem fairly stable. Since I've never actually
run qtile on pypy*, I can't say if it even actually works, so it seems best not
to gate commits on these.