Extend prompt widget for optional command history.
I missed this feature, usual in other aplication launchers
(like gmrun or the integrated in awesome WM, my previous
favorite tiling), so I wrote it :).
After thinking a bit about it, I concluded that the best way
was to use a deque for store the history, and serialize to a file
with pickle. Using plain text seemed more complicated and less
optimal to me.
When I changed handle_KeyPress first time, flake8 complained
(correctly) that the function was too much complex, so I
changed the implementation, dividing the function in a bunch
of little functions.