Inline command edition and a few errors subsaned.
Added support for cursor movement across the input text,
edition in place (add/delete text before the cursor, and
delete in cursor position with Delete key).

Added options for show or not the text cursor and for
choose their color. Cursor position is noticeable by
the text cursor and by the letter above it in the same
color than the cursor (even if you choose to not show the
text cursor).

Corrected a few errors in previous code:
- When cursorblink was set to zero, the cursor simply
dissapear. If this is a feature, is deceiving: If
cursorblink is zero it should, you know, just not
blink. That's what it do now.
- The widget didn't clear the text in focus change with
cursorblink with zero value. After my changes, this also
happened when you choose not show the cursor. Now it do in both.

Also, function in charge of return the appropiate action to one
key (_key_handler) was renamed (_get_keyhandler) and simplified,
creating a dictionary on init (for no rebuilt it every time a new
key is handled), with pairs (keycode, function to execute).