Add compatibility hacks for PyPy < 2.6
PyPy has cffi built-in, and it cannot be upgraded.  The current version
of PyPy (2.6.0) ships with cffi 1.1.0, but Travis still runs with PyPy
2.5.0, which has cffi 0.9, and so the features in cffi 1.0 cannot be
used.  This adds a compatibility layer for cffi 0.9, and so this can be
reverted when PyPy 2.6 is used on Travis and when support for PyPY <2.6
is dropped.