Fix version-dependent dependencies
Does anyone even use 3.0 or 3.1? If they do, this dependency check is
broken, the first if/elif/else block is solely for checking the event
loop requirements, make the importlib a separate check.

Looking at this, it would be nice if we could make the 3.3 check for
either trollius or asyncio, rather than enforcing asyncio, but I'm not
sure we can get setuptools to do that.