rework the focus model a bit
Several issues here:

1. most layouts were passing warp=False in a lot of places, so people who
   used cursor_warp effectively couldn't use those layouts. Instead, let's
   only pass warp=False where it makes sense

2. toScreen defaulted warp=True, which happened on EnterNotify, so the
   mouse would move around if you were switching screens with it, which was
   annoying.

Instead, let's make warp optional, and explicitly pass False when we know
we don't want to warp. This patch leaves max (and similar layouts) as they
were, not ever moving the pointer on behalf of the user, since is no real
"movement" when switching windows in these layouts.

Closes #760