don't crash when loading history
For whatever reason, my qtile history couldn't be loaded across restarts,
which causes this code to fail, which then causes me to lose my session.
Instead, let's just log an error (which, unfortuantely only goes to stderr
because this likely happens in the user's config before qtile's logging is
initialized) and continue on.