Pin requirements with pip-tools
Splits production and development packages in seperate requirements files.

In requirements.txt includes only the packages what are used for a
production enviroment. In requirements-dev.txt are production and development
dependencies unpinned to test on travis for the latest packages if something
breaks.

This commit adds also pip-tools as development tool to pin and sync the
requirements with the local enviroment.

Updates Makefile to add a new command to compile requirements.in to pin
the latest packages in the requirements.txt

@flacjacket propose this  way in qtile/qtile#779