Adds tox support
This add isloated virtualenvs for testing the code so it get easy to
test this code locally to test to the specific python version.
The virtualenvs populates only the needed requirements

Pep8 check are excluded to run only once on python 2 and 3.
Also this commit include a testenv to build documentation and one to
check the manifest and readme files.

Flake8 config move to the new setup.cfg to have these settings
in the root project
Requirements-dev.txt got a cleaup because this packages are
already define in the requirements.txt/in

Add nightly as allow_failure in the travis.yml