Adding support for CPU graph in FreeBSD
This commit adds support for CPUGraph in FreeBSD. And to make tests
complete in FreeBSD I also added support for the MemoryGraph.

There are other means of getting cpu and mem stats, but this is the
most similar to the vanilla graph.

For this to work in FreeBSD linprocfs needs to be mounted.

Since linprocfs uses an older model of the kernel not all
fields are available, hence removing some of the fields
for FreeBSD.

In FreeBSD /proc stuff is located in
/compat/linux/proc , but it lacks some of the features
of a modern linux kernel. Hence the platform.system(checks).

This should maintain compatibility with Linux, but please test
it so I don't break anything.

Cheers!

Closes #796