Fix make check command
'make check' don't find tests. This was introuced in
728fafdc937b56319d24d87bb21bf2f254dda92a

The problem is that the 'libqtile' was the argument of --cov.
That one restricted the coverage output. After this parameter was
removed
nosetests was thinking it should search in the folder libqtile for
tests.

I remove 'libqtile' as argument and now it runs smoothly again.
Also, I expand the command line parameter to see what they
actually mean and the 'make check-cov' call has now an echo output.