Fix the mocking the __getattr__ function
First, we don't actually need to install mock for Python 3.3, lets use
the version in the standard library.

We shouldn't be implementing our own version of MagickMock.__getattr__,
as it does some magic for us [1]. The only thing that we need this for is
cffi.__version_info__, but we can mock the ffi packages so the cffi
version does not need to be checked.

Finally, fix the creation of lists for sphinx.

[1]
https://docs.python.org/3/library/unittest.mock.html#mocking-magic-methods