Add CONTRIBUTING.md
In the README was the direct link missing to the contributing section in
the docs.

I add the Link to the docs about contributing.
Also to improve contributions I include the CONTRIBUTING.md this is file what
provides on github a yellow box what links contributor to the CONTRIBUTING.md.

https://help.github.com/articles/setting-guidelines-for-repository-contributors/

Closes #804

Closes #808