debian: update packaging
* add debian WNPP bug number
* bump versions for dependencies
* bump version number for release (this may need to be corrected)
* bump debian standards version (lintian)
* get rid of separate licensing for debian/ (lintian)
* add extra dependencies for widgets (still missing are google calendar's deps,
  iwlib, and mpd in python3 which aren't in the debian archives; if users
  install these manually they should work)

Note that there is still one (set of) lintian warning(s, for all the images),
which would require code changes to fix:

W: python-qtile: image-file-in-usr-lib usr/lib/python2.7/dist-packages/libqtile/resources/battery-icons/battery-good.png

I've also bumped the version number to 0.10.3, but that may not be the next
version of qtile that is tagged. We should bump this version number to whatever
that next version is when it is tagged. (And it would be handy to tag soon, so
I can start the process of getting a debian upload :)