debian: switch library packages to Architecture: any
Architecture: all means that the packages will work on all architectures,
which is not the case since libqtile now uses cffi to build C extensions;
these must be built per-arch. However, the qtile binary package contains
only man pages, the .desktop, and the qtile* binaries which are written in
pure python, so it can stay Architecture: all.