make: update deb building command
For whatever reason, the debian guys got rid of the git alias for this by
default, so now on a fresh vm you need to use the actual `gbp` binary instead.

Also, let's always build the package on the current branch (which means that we
need to remove the orig.tar.gz and friends in .. when building a new package).