Add iqsh kernel
Add Jupyter kernel that connects to qsh. Implement basic execution and
code completion for the kernel.

To install the kernel, run:

    python -m libqtile.interactive.iqsh_install

To run the kernel in a standalone instance, you can run:

    python -m libqtile.interactive.iqsh_kernel

To it will likely be easier to run spin up a kernel and run straight
from a Jupyter console:

    jupyter console --kernel iqsh

Or run the helper scripts (which just executes the above), `iqsh`.