Fixes and added docs for treetab
Fixed up some behavior such that it matches its intent.  Also, fix a
problem with Python 3 (and maybe even Python 2), as `six.u()` would not
return a string.

Unfortunately, TreeTab is totally untested, so there is no way to ensure
this isn't breaking things, but it was definitely broken before.