Handle invalid time on colored progess formatting
The total time can be 0 - i.e. in case of an audio stream, which raises a
ZeroDivisionError exception when formatting the text for the MPD Widget.
To prevent that - and maybe other issues - it's moved to a try/except block
to catch all exceptions. It's maybe "overkill" but prevent the widget to fail.