Stop cmus widget flooding stdout
The widget, since always, has been flooding the stdout with a
«cmus-remote: cmus is not running» message when cmus is not
running.

This behaviour doesn't affect the proper operation of the widget, nor
get logged in any place, so I has been delaying the fix to
«when I have some time».

It just bugged me a bit every time I shutdown qtile, because then
you saw a great sequence of that message filling your screen.
So I fixed it.