Refactor common code of layouts and fix window focus cycle.
Distribute common code of layouts to a new base class in layouts/base.py.
This class is used as base for layouts maintaining a plain list of clients
and a current client; used also for Columns and Stack layouts.

This fixes window focus cycle of Group for Columns and Stack layouts.

The Group expects layouts to cycle _all_ tiled windows.
If the end of managed windows is passed, None must be returned
in order to let group also cycle windows in the floating layer.
Docstrings are added to document this relationship.

For wmii only the window cycle is fixed, without rework.