update btc tracker to use coinbase
Seeems btc-e has died, so that was a bad choice. Coinbase is much more
reputable, perhaps they'll stick around for a few more years.

Note that this drops quite a bit of the flexibility of the previous widget
w.r.t. other currencies, but adds support for many more fiat currencies.
This is simply because of the exchange capabilities. However, I've taken
care to make sure that old configs will at least not crash the new widget,
with a best-effort to honor the ones that can be honored.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>