conftest: handle errors in qtile constructor
The problem here is that can_connect_qtile() used to raise an
AssertionError instead of returning False, so the subsequent handling code
would never be called, and it was hard to determine the cause of the
problem.

There is a similar issue with can_connect_x11(), so let's handle that too.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>