Start refactoring to isolate core
This commit moves towards refactoring qtile to extract the core.
Eventually, all code that interacts with X will be isolated in the core,
and a well-defined interface will determine how the core interacts with
the rest of the codebase.

This is just a first step, placing the relevant files in the core
directory, and adjusting import paths. I also shift the xcursors CFFI
bindings into core, since they are only used by the core module. This
means splitting the ffi_build script, and providing a convenience script
in ./scripts to build both ffi modules.