Shift event reading to core
Event reading for back-ends is now located within the Core. Events are
passed to the manager on a queue, and handled asynchronously within the
loop.

One complication with this patch is that it exacerbates the existing
raciness of our test suite. Many tests incorrectly rely on synchronous
behaviour, and don't succeed reliably. With this patch, these tests are
more likely to fail, especially on faster systems. Improving test
reliability will require test infrastructure that will be made possible
by the core refactoring.