copy default config attributes over to test configs
The recent cursor_warp fix meant that now some of the test code configs
need cursor_warp. It's a pain in the ass to add unrelated variables to the
test configs every time we add a new config variable, and it trips a lot of
people up.

Instead, let's just copy everything that doesn't already exist over from
the default config, so nobody has to mess around with this any more.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>