don't wait for 10 seconds after SIGINT
For whatever reason qtile doesn't respond to SIGINT any more (at least on
my machine), so this causes an unconditional sleep for 10 seconds when I
run the tests.

Given that we don't really need this thing to shut down nicely, let's just
SIGKILL it and be done. No need to be polite.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>