Make navigating command graph explicit
Change from navigating the command graph via magic methods to using an
explitcit navigation command.  This makes type checking and inference
easier, particularly when the command graph is used in the interactive
and lazy clients.  Get rid of the separate command graph command that is
generated by the __call__ magic, we can directly use the command graph
call object to create this as needed.