Create new lazy graph structure
Create a new lazy command graph structure, deprecating the previous one.
This gives a better separation between the object graph itself and the
commands that are executed on the object graph.  The lazy objects can
now also be decorated with custom logic for checks, rather than having
the logic have to be built into the lazy object.