tests: drop need for xvfbwrapper
this is mostly redundant since the xcffib test suite (with a little
modification) contains its own implementation of this, so let's drop it.

as a bonus, when manually allocating the display (for use with xephyr) we
can get rid of this warning:

    test/test_hook.py::test_can_call_by_selection_notify
      /home/tycho/packages/qtile/test/conftest.py:195: ResourceWarning: unclosed file <_io.TextIOWrapper name='/tmp/.X358037035-lock' mode='w' encoding='UTF-8'>
        self.display = ":{}".format(_find_display())

since we won't lose the fd, since xcffib doesn't keep it around. we do need
to delete the file when we're done though, so let's do that.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>