Fix a weird length calculation issue on textboxes
I thought it was smart to only assign the text to the layout when it's
drawn (see https://github.com/qtile/qtile/pull/1367/files#r300469727),
but it seems that it's used for the calculation of the widget's length
beforehand and result in buggy behaviors that way (either there is a delay
before the widget is drawn at the proper size, or it is just drawn with a wrong
length and stays that way). So let's revert to a property setter method; plus
it will give a better output for info().