docs: Explain better how the Xephyr script works
The hacking.rst file is reorganized to present information
in the best order for a new contributor.

First, the requirements: Requirements and Building cffi module.

Second, an overview of how to develop and test: Development
and testing, Coding style and Deprecation policy.

Then, a more detailed explanation of how the tests work
(with Xephyr), and how to use the Xephyr helper script.

Finally, more advanced things, like using a second X session,
or capturing an xtrace.