Fix test_images2 test
Drop PSNR from the list of metrics that are tested, as all of the others
are best at low values, while PSNR is best at high values. It seems like
the scaling is basically perfect, which is expected for SVG's, and so
the PSNR blows-up to inf.

Lower the imagemagick version requirement to run the test so that it
runs on Travis. It is able to pass here, so I'm not sure where this
requirument is coming from