Directly import the currentlayout widgets
There is currently a mypy failure in the default layout because it is
importing one of these widgets and it is not able to introspect through
the safe import to see the call.  The currentlayout widgets do not use
any third party libraries, so it is safe to directly import these at the
top level.  Add the widget imports here, as well as a note on what can
and cannot be included via direct import.