confreader: drop bad default keys/mouse settings
72f35e2c2128 ("Re-enable Python 3.5 support") added these lines, which
break our default resolution by causing them to be present on the current
object, so the defaults from the default config will not be read.

Instead, let's just drop these and have them be set normally as everything
else is in the default config options.

Thanks to @actbusy for debugging and getting most of the way there in #1411

Signed-off-by: Tycho Andersen <tycho@tycho.ws>