Makefile: some basic cleanups
We've migrated to tox for linting/testing, and now that we do a bunch of
additional linting on imports, etc., we should probably use the same thing
for the make targets that we do for CI. (Am I the only one that uses the
Makefile? :D)

We can drop ckpatch, as that was originally the entry point for travis. I
just tend to run individual things manually.

Also, let's clean up some of the new coverage reports that tox generates.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>