docs: call setuptools_scm directly to get version string
This is explicitly advised against in the documentation [1] but
installing qtile as a dependency from the docs is not an option. Using
pkg_resources is therefore not working.

The dependency chain for qtile is too complex for pip, since cairocffi
must be installed prior to executing qtile/setup.py but there's not
way to sequence this correctly in qtile/requirements.txt.

Another option is to override sphinx's version string in
qtile/setup.py using command_options to 'build_sphinx' [2] but this
doesn't allow building the docs using the Makefile in docs/ anymore.

[1] https://github.com/pypa/setuptools_scm#usage-from-sphinx
[2] https://www.sphinx-doc.org/en/master/usage/advanced/setuptools.html

Signed-off-by: Etienne Wodey <airwoodix@posteo.me>